Fixup

+++ shell.c
@@ -1,51 +1,49 @@
 #include <signal.h>
 #include <stdio.h>
+#include <stdbool.h>
 #include <stdlib.h>
-#include <termios.h>
+#include <string.h>
 #include <unistd.h>
+#include <termios.h>
 
-#include "io.h"
 #include "shell.h"
+#include "io.h"
 
+/* External variables declared at shell.h. Their purpose is explained
+   at shell.h*/
 int shell_terminal;
 bool shell_is_interactive;
 pid_t shell_pgid;
 struct termios shell_tmodes;
 
-/**
- * Setting standard I/O file descriptors to correct values.
- */
+/* Set appropriate variables to associate standard input file descriptor
+   to foreground terminal process and killing all other background terminal
+   proceses.*/
 void initialize_shell()
 {
-  /* Checking whether standard input is associated with a */
-  /* foreground terminal */
   shell_terminal = STDIN_FILENO;
   shell_is_interactive = isatty(shell_terminal);
 
-  /* Kill all process groups not associated with shell terminal */
-  /* file descriptor */
-  while (tcgetpgrp(shell_terminal) != (shell_pgid = getpgrp())) {
-    kill(shell_pgid, SIGTTIN);
+  if (shell_is_interactive) {
+    /* Kill all background processes. */
+    while (tcgetpgrp(shell_terminal) != (shell_pgid = getpgrp())) {
+      kill(shell_pgid, SIGTTIN);
+    }
+
+    /* Bring terminal process to foreground. */
+    shell_pgid = getpgrp();
+    tcsetpgrp(shell_terminal, shell_pgid);
+    tcgetattr(shell_terminal, &shell_tmodes);
   }
-
-  /* Process group id of foreground process */
-  shell_pgid = getpgrp();
-
-  /* Put standard input file descriptor to foreground */
-  tcsetpgrp(shell_terminal, shell_pgid);
-  tcgetattr(shell_terminal, &shell_tmodes);
 }
 
-/**
- * Core shell logic
- */
 int shell(int argc, char* argv[])
 {
   char *command_string;
 
   initialize_shell();
 
-  while ((command_string = get_next_command_string())) {
+  while ((command_string = get_next_command_string()) != NULL) {
     printf("%s", command_string);
   }
   return 0;
